Dropped the python wrapper header to simplify the packaging
