+++ Hector/PythonWrapper.cc
@@ -1,5 +1,3 @@
-#include "Hector/PythonWrapper.h"
-
 #include "Hector/Core/ExceptionType.h"
 #include "Hector/Core/Parameters.h"
 
@@ -16,11 +14,15 @@
 
 #include "Hector/IO/MADXHandler.h"
 
+#include <boost/python.hpp>
+
 #include <memory>
 #include <map>
 
 namespace
 {
+  namespace py = boost::python;
+
   std::string dump_particle( const Hector::Particle& part ) {
     std::ostringstream os;
     part.dump( os );
@@ -46,6 +48,22 @@ namespace
   }
   py::list beamline_elements( Hector::Beamline& bl ) {
     return to_python_list<std::shared_ptr<Hector::Element::ElementBase> >( bl.elements() );
+  }
+
+  struct ElementBaseWrap : Hector::Element::ElementBase, py::wrapper<Hector::Element::ElementBase>
+  {
+    ElementBaseWrap() : Hector::Element::ElementBase( Hector::Element::anInvalidElement ) {}
+    std::shared_ptr<Hector::Element::ElementBase> clone() const override { return this->get_override( "clone" )(); }
+    Hector::Matrix matrix( float eloss, float mp, int qp ) const override { return this->get_override( "matrix" )( eloss, mp, qp ); }
+  };
+
+  template<class T, class init = py::init<std::string,float,float,float> >
+  void convertElement( const char* name )
+  {
+    py::class_<T, py::bases<Hector::Element::ElementBase> >( name, init() )
+      .def( "clone", &T::clone, py::return_value_policy<py::return_by_value>() )
+      .def( "matrix", &T::matrix )
+    ;
   }
 }
